refactor: modifing project structure
- Jungle 내부 구조 수정 진행중
- Carousel MVC 구조로 수정 진행중 (CardModel 작성)
- 각 UI 테스트용 page 폴더 생성
