+++ Jungle/components/Carousel/Models/CardModel.js
@@ -1,0 +1,9 @@
+export default class CardModel {
+  constructor() {
+    this.carouselItems = {};
+  }
+
+  updateCarouselItem(updatedItems) {
+    this.carouselItems = {...this.carouselItems, ...updatedItems};
+  }
+}
