+++ data/localData.json
@@ -5,7 +5,7 @@
     "desc": [
       "This is a First card."
     ],
-    "imgUrl": "./images/Card_A01.png"
+    "imgUrl": "../images/Card_A01.png"
   },
   {
     "id": 2,
@@ -13,7 +13,7 @@
     "desc": [
       "This is a Second card."
     ],
-    "imgUrl": "./images/Card_B01.png"
+    "imgUrl": "../images/Card_B01.png"
   },
   {
     "id": 3,
@@ -21,7 +21,7 @@
     "desc": [
       "This is a Third card."
     ],
-    "imgUrl": "./images/Card_C01.png"
+    "imgUrl": "../images/Card_C01.png"
   },
   {
     "id": 4,
@@ -29,6 +29,6 @@
     "desc": [
       "This is a Fourth card."
     ],
-    "imgUrl": "./images/Card_D01.png"
+    "imgUrl": "../images/Card_D01.png"
   }
 ]
