+++ pages/carousel.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+  <title>Carousel</title>
+  <link rel="stylesheet" href="../css/style.css" />
+</head>
+
+<body>
+  <h2>결합동작 예시</h2>
+  <div class="nav">
+    <div class="nav-item steelblue">Fast, Free Delivery</div>
+    <div class="nav-item purple">Exclusive deals and offers</div>
+    <div class="nav-item darkgreen">
+      Prime Originals, movies and TV shows
+    </div>
+    <div class="nav-item orange">Over 2 million songs ad free</div>
+  </div>
+  <div class="container">
+
+    <div class="card-wrapper">
+      <div class="card-slider">
+
+      </div>
+    </div>
+    <button class="card-control prev">
+      <</button> <button class="card-control next">>
+    </button>
+  </div>
+
+  <h2>개별동작 예시 (infinite : off)</h2>
+  <div class="nav2">
+    <div class="nav-item steelblue">1</div>
+    <div class="nav-item purple">2</div>
+    <div class="nav-item darkgreen">
+      3
+    </div>
+    <div class="nav-item orange">4</div>
+  </div>
+  <div class="container2">
+    <div class="card-wrapper">
+      <div class="card-slider2">
+      </div>
+    </div>
+    <button class="card-control prev">
+      <</button> <button class="card-control next">>
+    </button>
+  </div>
+  <script type="module" src="../js/carouselApp.js"></script>
+</body>
+
+</html>
